+++ lib/SecondTap.dart
@@ -1,4 +1,13 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:team7_work/FirstTab.dart';
+import 'package:provider/provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'main.dart';
+import 'FirstTab.dart';
+import 'memo_service.dart';
 
 class SecondTab extends StatelessWidget {
   const SecondTab({Key? key}) : super(key: key);
