+++ lib/main.dart
@@ -10,6 +10,7 @@ import 'package:team7_work/FifthTab.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'memo_service.dart';
 
 late SharedPreferences prefs;
